% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chol_stable}
\alias{chol_stable}
\title{chol_stable}
\usage{
chol_stable(X)
}
\arguments{
\item{X}{data set or matrix}
}
\value{
cholesky-decomposed matrix
}
\description{
numerically stable Cholesky decomposition
}
