% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getb_maxvar}
\alias{getb_maxvar}
\title{getb_maxvar}
\usage{
getb_maxvar(X)
}
\arguments{
\item{X}{data set or matrix}
}
\value{
cholesky-decomposed matrix
}
\description{
numerically stable Cholesky decomposition
}
