% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_rdd_plot}
\alias{gp_rdd_plot}
\title{gp_rdd_plot}
\usage{
gp_rdd_plot(gp_rdd_res, l_col = "blue", r_col = "red")
}
\arguments{
\item{gp_rdd_res}{a list-form results obtained from gp_rdd()}

\item{l_col}{a character value indicating the color of the left side of the cutoff point (default = "blue")}

\item{r_col}{a character value indicating the color of the right side of the cutoff point (default = "red")}
}
\value{
\item{gg}{an RD ggplot}
}
\description{
to draw an RD plot using the results obtained from gp_rdd()
}
