% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_train}
\alias{gp_train}
\title{gp_train}
\usage{
gp_train(
  X,
  Y,
  b = NULL,
  s2 = 0.3,
  optimize = FALSE,
  scale = TRUE,
  mixed_data = FALSE,
  cat_columns = NULL,
  Xtest = NULL
)
}
\arguments{
\item{X}{a set of covariate data frame or matrix}

\item{Y}{Y vector (outcome variable)}

\item{b}{bandwidth (default = NULL)}

\item{s2}{noise or a fraction of Y not explained by X (default = 0.3)}

\item{optimize}{a logical value to indicate whether an automatic optimized value of S2 should be used. If FALSE, users must define s2. (default = FALSE)}

\item{scale}{a logical value to indicate whether covariates should be scaled. (dafault = TRUE)}

\item{mixed_data}{a logical value to indicate whether the covariates contain a categorical/binary variable (default = FALSE)}

\item{cat_columns}{a character or a numerical vector indicating categorical variables (default = NULL)}

\item{Xtest}{a data frame or a matrix of testing covariates. This is necessary when a non-overlapping categorical value exists between training and testing data sets. (default = NULL)}
}
\value{
\item{post_mean_scaled}{posterior distribution of Y in a scaled form}
\item{post_mean_orig}{posterior distribution of Y in an original scale}
\item{post_cov_scaled}{posterior covariance matrix in a scaled form}
\item{post_cov_orig}{posterior covariance matrix in an original scale}
}
\description{
Feed training data set to gp_train() to obtain a gp model.
}
