% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gpss}
\alias{predict.gpss}
\title{predict method for \code{gpss} objects}
\usage{
\method{predict}{gpss}(
  object,
  newdata = NULL,
  type = "response",
  format = "default",
  interval = "confidence",
  level = 0.95,
  prior_mean = NULL,
  ...
)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{data frame on which to make predictions (test set)}

\item{type}{"response" or "scaled"}

\item{format}{"default" or "rvar"}

\item{interval}{"prediction" or "confidence"}

\item{level}{a numerical value between 0 and 1}

\item{prior_mean}{a numeric vector of prior mean values for Y at each test observation. Required when the model was trained with a \code{prior_mean}; see \code{\link{gp_predict}}. (default = NULL)}

\item{...}{additional arguments (not used)}
}
\description{
predict method for \code{gpss} objects
}
\examples{
library(gpss)
data(lalonde)
# categorical variables must be encoded as factors
dat <- transform(lalonde, race_ethnicity = factor(race_ethnicity))
# train and test sets
idx <- sample(seq_len(nrow(dat)), 500)
dat_train <- dat[idx, ]
dat_test <- dat[-idx, ]
# sample of data for speed
mod <- gpss(re78 ~ nsw + age + educ + race_ethnicity, data = dat_train)
p <- predict(mod, newdata = dat_test)
p_confidence99 <- predict(mod, newdata = dat_test, interval = "confidence", level = 0.99)
}
