% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_its.R
\name{print.gp_its}
\alias{print.gp_its}
\title{Print method for gp_its objects}
\usage{
\method{print}{gp_its}(x, ...)
}
\arguments{
\item{x}{a gp_its object}

\item{...}{additional arguments (not used)}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for gp_its objects
}
\keyword{internal}
