% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_rdd.R
\name{print.gp_rdd}
\alias{print.gp_rdd}
\title{Print method for gp_rdd objects}
\usage{
\method{print}{gp_rdd}(x, ...)
}
\arguments{
\item{x}{a gp_rdd object}

\item{...}{additional arguments (not used)}
}
\description{
Print method for gp_rdd objects
}
\keyword{internal}
