% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_its.R
\name{summary.gp_its}
\alias{summary.gp_its}
\title{Summary method for gp_its objects}
\usage{
\method{summary}{gp_its}(object, ...)
}
\arguments{
\item{object}{a gp_its object}

\item{...}{additional arguments (not used)}
}
\description{
Summary method for gp_its objects
}
\keyword{internal}
