% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_rdd.R
\name{summary.gp_rdd}
\alias{summary.gp_rdd}
\title{Summary method for gp_rdd objects}
\usage{
\method{summary}{gp_rdd}(object, ...)
}
\arguments{
\item{object}{a gp_rdd object}

\item{...}{additional arguments (not used)}
}
\description{
Summary method for gp_rdd objects
}
\keyword{internal}
