% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_predict}
\alias{gp_predict}
\title{gp_predict}
\usage{
gp_predict(gp, Xtest)
}
\arguments{
\item{gp}{a list-form object obtained from gp_train()}

\item{Xtest}{a data frame or a matrix of testing data set}
}
\value{
\item{Xtest_scaled}{scaled testing data set}
\item{Xtest}{original testing data set}
}
\description{
gp_optimize() optimizes S2 using MLE. This function automatically runs when gp_train(..., optimize=TRUE).
}
