% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpss-package.R, R/gpss.R
\docType{package}
\name{gpss}
\alias{gpss-package}
\alias{gpss}
\title{gpss: Gaussian Processes in Social Science}
\usage{
gpss(formula, data, b = NULL, s2 = 0.3, optimize = FALSE, scale = TRUE)
}
\arguments{
\item{b}{bandwidth (default = NULL)}

\item{s2}{noise or a fraction of Y not explained by X (default = 0.3)}

\item{optimize}{a logical value to indicate whether an automatic optimized value of S2 should be used. If FALSE, users must define s2. (default = FALSE)}

\item{scale}{a logical value to indicate whether covariates should be scaled. (dafault = TRUE)}
}
\value{
\item{post_mean_scaled}{posterior distribution of Y in a scaled form}
\item{post_mean_orig}{posterior distribution of Y in an original scale}
\item{post_cov_scaled}{posterior covariance matrix in a scaled form}
\item{post_cov_orig}{posterior covariance matrix in an original scale}
\item{K}{a kernel matrix of X}
\item{prior_mean_scaled}{prior distribution of mean in a scaled form}
\item{X.orig}{the original matrix or data set of X}
\item{X.init}{the original matrix or data set of X with categorical variables in an expanded form}
\item{X.init.mean}{the initial mean values of X}
\item{X.init.sd}{the initial standard deviation values of X}
\item{Y.init.mean}{the initial mean value of Y}
\item{Y.init.sd}{the initial standard deviation value of Y}
\item{K}{the kernel matrix of X}
\item{Y}{scaled Y}
\item{X}{scaled X}
\item{b}{bandwidth}
\item{s2}{sigma squared}
\item{alpha}{alpha value in Rasmussen and Williams (2006) p.19}
\item{L}{L value in Rasmussen and Williams (2006) p.19}
\item{mixed_data}{a logical value indicating whether X contains a categorical/binary variable}
\item{cat_columns}{a character or a numerical vector indicating the location of categorical/binary variables in X}
\item{cat_num}{a numerical vector indicating the location of categorical/binary variables in an expanded version of X}
\item{Xcolnames}{column names of X}
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Gaussian Process (GP) combines a highly flexible non-linear regression approach with rigorous handling of uncertainty. A key feature of this approach is that, while it can be used to produce a conditional expectation function representing the mode of a posterior conditional distribution, it does not ``choose a particular model fit'' and construct uncertainty estimates conditional on putting full faith in that model. This is valuable because the uncertainty estimates reflect the lesser knowledge we have at locations near, at, or beyond the edge of the observed data, where results from other approaches would become highly model-dependent. We first offer an accessible explanation of GPs, and provide an implementation more suitable to social science inference problems, which reduces the number of user-chosen hyperparameters from three to zero. We then illustrate the settings in which GPs can be most valuable: those where conventional approaches have poor properties due to model-dependency/extrapolation in data-sparse regions. Specifically, we demonstrate the usefulness of GPs in contexts where (i) treated and control models are needed by these groups have poor covariate overlap; (ii) regression discontinuity, which depends on model estimates taken at or just beyond the edge of their supporting data; and (iii) interrupted time-series designs, where models are fitted prior to an event by extrapolated after it.
}
\examples{
library(gpss)
data(lalonde)

# categorical variables must be encoded as factors
dat <- transform(lalonde, race_ethnicity = factor(race_ethnicity))

# train and test sets
idx <- sample(seq_len(nrow(dat)), 500)
dat_train <- dat[idx, ]
dat_test <- dat[-idx, ]

# Fit model
mod <- gpss(re78 ~ nsw + age + educ + race_ethnicity, data = dat_train)

# predictions in the test set
p <- predict(mod, dat_test)
length(p)
head(p)
}
\seealso{
Useful links:
\itemize{
  \item \url{http://doeunkim.org/gpss/}
  \item Report bugs at \url{https://github.com/doeun-kim/gpss/issues}
}

}
\author{
\strong{Maintainer}: Chad Hazlett \email{chazlett@ucla.edu}

Authors:
\itemize{
  \item Soonhong Cho \email{tnsehdtm@gmail.com}
  \item Doeun Kim \email{doeun2@ucla.edu} (\href{https://orcid.org/0000-0003-4789-6599}{ORCID})
}

}
\keyword{internal}
