% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpss-package.R
\docType{package}
\name{gpss}
\alias{gpss}
\alias{gpss-package}
\title{gpss: Gaussian Processes in Social Science}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Gaussian Process (GP) combines a highly flexible non-linear regression approach with rigorous handling of uncertainty. A key feature of this approach is that, while it can be used to produce a conditional expectation function representing the mode of a posterior conditional distribution, it does not ``choose a particular model fit'' and construct uncertainty estimates conditional on putting full faith in that model. This is valuable because the uncertainty estimates reflect the lesser knowledge we have at locations near, at, or beyond the edge of the observed data, where results from other approaches would become highly model-dependent. We first offer an accessible explanation of GPs, and provide an implementation more suitable to social science inference problems, which reduces the number of user-chosen hyperparameters from three to zero. We then illustrate the settings in which GPs can be most valuable: those where conventional approaches have poor properties due to model-dependency/extrapolation in data-sparse regions. Specifically, we demonstrate the usefulness of GPs in contexts where (i) treated and control models are needed by these groups have poor covariate overlap; (ii) regression discontinuity, which depends on model estimates taken at or just beyond the edge of their supporting data; and (iii) interrupted time-series designs, where models are fitted prior to an event by extrapolated after it.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://doeunkim.org/gpss/}
  \item Report bugs at \url{https://github.com/doeun-kim/gpss/issues}
}

}
\author{
\strong{Maintainer}: Chad Hazlett \email{chazlett@ucla.edu}

Authors:
\itemize{
  \item Soonhong Cho \email{tnsehdtm@gmail.com}
  \item Doeun Kim \email{doeun2@ucla.edu} (\href{https://orcid.org/0000-0003-4789-6599}{ORCID})
}

}
\keyword{internal}
