% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gpss}
\alias{predict.gpss}
\title{predict method for \code{gpss} objects}
\usage{
\method{predict}{gpss}(object, newdata = NULL, type = "response")
}
\description{
predict method for \code{gpss} objects
}
